Created a separate file for storing application configurations and also created a file that has all the email confirmation stuff in it
