+++ nightbus-app/config.py
@@ -1,0 +1,10 @@
+class TestConfig:
+    DEBUG = True
+    TESTING = False
+    CSRF_ENABLED = True
+    SECRET_KEY = 'This is secret'
+    MAIL_SERVER = 'smtp.gmail.com'
+    MAIL_PORT = 465
+    MAIL_USE_SSL = True
+    MAIL_USERNAME = 'abenezer.mammo@gmail.com'
+
